+++ CHANGELOG.md
@@ -87,6 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `internal/xworkspaces`: `%nwin%` can be used to display the number of open
   windows per workspace
   ([`#604`](https://github.com/polybar/polybar/issues/604))
+- `internal/backlight`: added `use-actual-brightness` option
 
 ### Changed
 - Slight changes to the value ranges the different ramp levels are responsible
@@ -106,6 +107,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([`#2292`](https://github.com/polybar/polybar/issues/2292))
 - Parser error if click command contained `}`
   ([`#2040`](https://github.com/polybar/polybar/issues/2040))
+- `internal/backlight`: With amdgpu backlights, the brightness indicator was slightly behind.
+  ([`#2367](https://github.com/polybar/polybar/issues/2367))
 
 ## [3.5.4] - 2021-01-07
 ### Fixed